rearranging guts
