+++ app/root.jsx
@@ -1,6 +1,6 @@
 import { redirect } from '@remix-run/node';
 import { Links, Meta, Outlet, Scripts, ScrollRestoration, useLoaderData, useRouteError } from '@remix-run/react';
-import { getSession } from './session.server';
+import { getSession } from './utils/session.server';
 import { UserProvider } from './context/UserContext';
 import globalStyles from './styles/global.css?url';
 
@@ -11,54 +11,17 @@ export const links = () => [
 ];
 
 export const loader = async ({ request }) => {
-	const url = new URL(request.url);
-	const path = url.pathname;
-	const shop = process.env.SHOPIFY_STORE_DOMAIN;
 	const session = await getSession(request.headers.get('Cookie'));
-	console.log('🛩️ Navigating to page: ', path);
+	const url = new URL(request.url);
+	const userData = session.get('userData');
+	const shop = process.env.SHOPIFY_STORE_DOMAIN;
 
-	const customerAccessToken = session.get('customerAccessToken');
-	if (!customerAccessToken && path !== '/gsan/login') {
-		console.log('🏓 Missing user session. Redirecting to /gsan/login');
-		return redirect(`/gsan/login?shop=${shop}`);
+	if (!userData && url.pathname !== '/auth') {
+		console.log('🚗 No user data, redirecting to /auth');
+		return redirect('/auth');
 	}
 
-	let user = null;
-	if (customerAccessToken) {
-		try {
-			const response = await fetch(`https://${shop}/api/2024-01/graphql.json`, {
-				method: 'POST',
-				headers: {
-					'Content-Type': 'application/json',
-					'X-Shopify-Storefront-Access-Token': process.env.SHOPIFY_STOREFRONT_ACCESS_TOKEN,
-				},
-				body: JSON.stringify({
-					query: `
-            query($customerAccessToken: String!) {
-              customer(customerAccessToken: $customerAccessToken) {
-                id
-                firstName
-                lastName
-                email
-              }
-            }
-          `,
-					variables: { customerAccessToken },
-				}),
-			});
-			const data = await response.json();
-			if (response.ok && data?.data?.customer) {
-				const { id, firstName, lastName, email } = data.data.customer;
-				user = { id, firstName, lastName, email };
-			} else {
-				console.error('Error fetching customer data:', data.errors || response.statusText);
-			}
-		} catch (error) {
-			console.error('Error during customer data fetch:', error);
-		}
-	}
-
-	return { shop, user };
+	return { userData, shop };
 };
 
 export default function Root() {
@@ -70,7 +33,10 @@ export default function Root() {
 				<Links />
 			</head>
 			<body>
-				<UserProvider value={user}>
+				<UserProvider
+					initialUser={user}
+					shop={shop}
+				>
 					<Outlet />
 				</UserProvider>
 				<ScrollRestoration />
