+++ app/routes/api.modem.jsx
@@ -1,6 +1,6 @@
 import { json } from '@remix-run/node';
 import axios from 'axios';
-import { getCompassAccessToken } from './api.get-compass-access-token';
+import { getCompassAccessToken } from '../compass.server';
 import fetchGPS from './api.gps';
 
 export const loader = async ({ params }) => {
