+++ app/utils/provider.server.js
@@ -1,0 +1,7 @@
+export async function getProviderCustomers(providerId) {
+	// Implement fetching customers for a specific provider
+}
+
+export async function getProviderMetafields(providerId) {
+	// Implement fetching provider-specific metafields
+}
